added back robots, refactor
